model: Add the class `LicenseInfoResolver`
Add the model for resolved license information and a related resolver
class. Resolved in this context means that all processing is done, for
example license finding curations are applied, and copyright statements
are processed.

The resolved license model contains all related data with the goal to be
completely transparent with regard to how license data was modified.
This information will later be needed by the reports to be able to show
to the user the effect of configuration options like license finding
curations.

The resolver class converts the raw license info into the resolved
license info and caches the results to avoid computing them repeatedly.
This commit only provides the basic implementation of the resolver, most
features need to be implemented in later commits. The initial
implementation also always returns the full data sets. This will later
be extended with filters to only get the required data, for example to
filter out license locations excluded by path excludes, or to get only
declared licenses.

At this point the implementation is completely separate from existing
license functions like the `LicenseResolver` or the license related
extension functions `OrtResultExtensions`. The plan is that all code in
ORT that works with licenses will gradually be migrated to the new API,
adding the required features to the API on the fly.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@here.com>