LicenseInfoResolver: Apply license finding curations
Apply license finding curations to the detected licenses. The
implementation currently has two shortcoming that need to be fixed
later. The first is that the original license findings are lost, which
would be required to show the effect of license finding curations in a
reporter, the other is that if multiple curations produce the same
result, only the first of the curations is recorded in the resolved
license.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@here.com>