ReportCommand: Create package configurations for projects
Create package configurations for the projects and pass them to the
license resolver. This allows the resolver to correctly handle path
excludes and license finding curations for projects.

Also add the project configurations to the package configuration
provider in the reporter input by default. This is reasonable default
behavior and useful for all reporter tests.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@here.com>